+++ app/Jobs/ProvisionDigitalOceanDropletJob.php
@@ -29,7 +29,9 @@ class ProvisionDigitalOceanDropletJob implements ShouldQueue
 
     public function __construct(
         public Server $server
-    ) {}
+    ) {
+        $this->onQueue(config('server_provision.queue', 'dply'));
+    }
 
     public function handle(): void
     {